+++ Assets/OxGKit/Utilities/Scripts/Samples~/CursorManagerDemo/Scripts/CursorManagerDemo.cs
@@ -1,4 +1,3 @@
-using OxGKit.Utilities.CursorAnim;
 using UnityEngine;
 using UnityEngine.InputSystem;
 using UnityEngine.UI;
@@ -12,16 +11,21 @@ public class CursorManagerDemo : MonoBehaviour
     private CursorLockMode _cursorLockMode;
     private int _cursorLockCount = 0;
 
+    private void Awake()
+    {
+        Cursors.InitInstance();
+    }
+
     private void Start()
     {
         // Initialize cursor visible = true
         this._cursorVisible = true;
-        CursorManager.GetInstance().SetCursorVisible(this._cursorVisible);
+        Cursors.SetCursorVisible(this._cursorVisible);
         this.cursorVisibleTxt.text = $"Visible: {this._cursorVisible}";
 
         // Initialize cursor lock mode = none
         this._cursorLockMode = CursorLockMode.None;
-        CursorManager.GetInstance().SetCursorLockState(this._cursorLockMode);
+        Cursors.SetCursorLockState(this._cursorLockMode);
         this.cursorLockModeTxt.text = $"Lock Mode: {this._cursorLockMode}";
     }
 
@@ -42,31 +46,31 @@ public class CursorManagerDemo : MonoBehaviour
     private void _SwitchCursorVisible()
     {
         this._cursorVisible = !this._cursorVisible;
-        CursorManager.GetInstance().SetCursorVisible(this._cursorVisible);
+        Cursors.SetCursorVisible(this._cursorVisible);
     }
 
     private void _CycleCursorLockMode()
     {
         this._cursorLockCount++;
         this._cursorLockMode = (CursorLockMode)(this._cursorLockCount % 3);
-        CursorManager.GetInstance().SetCursorLockState(this._cursorLockMode);
+        Cursors.SetCursorLockState(this._cursorLockMode);
     }
 
     public void SetDefaultCursor()
     {
         string stateName = "Default";
-        CursorManager.GetInstance().SetCursorState(stateName);
+        Cursors.SetCursorState(stateName);
     }
 
     public void SetHoverCursor()
     {
         string stateName = "Hover";
-        CursorManager.GetInstance().SetCursorState(stateName);
+        Cursors.SetCursorState(stateName);
     }
 
     public void SetLoadingCursor()
     {
         string stateName = "Loading";
-        CursorManager.GetInstance().SetCursorState(stateName);
+        Cursors.SetCursorState(stateName);
     }
 }